+++ src/app/(route)/votes/page.tsx
@@ -2,11 +2,15 @@ import { Suspense } from 'react'
 import MenuButtons from './_component/MenuButtons'
 import RankingList from './_component/RankingList'
 import VoteList from './_component/VoteList'
+import CreateVoteButton from './_component/CreateVoteButton'
 
 export default function page() {
   return (
     <main className="mx-auto w-[794px]">
-      <MenuButtons />
+      <div className="mt-[52px] flex justify-between">
+        <MenuButtons />
+        <CreateVoteButton />
+      </div>
       <Suspense fallback={<div>Loading. . . </div>}>
         <RankingList />
       </Suspense>
